+++ src/shared/backgroundScripts/actionButton.js
@@ -1,4 +1,4 @@
-import { isCurrentTabValidUrlScheme } from "./launchBrowser.js";
+import { isCurrentTabValidUrlScheme } from "./validTab.js";
 
 import { getDefaultIconPath, getGreyedIconPath } from "../../chromium/interfaces/getIcon.js";
 
@@ -7,7 +7,5 @@ export async function updateToolbarIcon() {
   if (!isCurrentTabValidUrlScheme) {
     iconPath = await getGreyedIconPath();
   }
-  
-  console.log("iconPath", iconPath);
   chrome.action.setIcon({ path: iconPath });
 }
