+++ src/shared/backgroundScripts/hotkeys.js
@@ -1,9 +1,9 @@
-import { launchFirefox } from "./launchBrowser.js";
+import { launchBrowser } from "../../chromium/interfaces/launchBrowser.js";
 
 export async function handleHotkeyPress(command, tab) {
   if (command === "launchFirefox") {
-    await launchFirefox(tab, true);
+    await launchBrowser(tab, true);
   } else if (command === "launchFirefoxPrivate") {
-    await launchFirefox(tab, false);
+    await launchBrowser(tab, false);
   }
 }
