--- src/shared/backgroundScripts/launchBrowser.js
@@ -1,35 +1,0 @@
-import { getIsFirefoxInstalled } from "./getters.js";
-
-export let isCurrentTabValidUrlScheme = false;
-
-export function checkAndUpdateURLScheme(tab) {
-  if (tab.url === undefined) isCurrentTabValidUrlScheme = false;
-  else if (tab.url.startsWith("http") || tab.url.startsWith("file")) {
-    isCurrentTabValidUrlScheme = true;
-  } else {
-    isCurrentTabValidUrlScheme = false;
-  }
-}
-
-export async function launchFirefox(tab, launchDefaultBrowsing) {
-  if (!(await getIsFirefoxInstalled())) {
-    chrome.tabs.create({ url: "https://www.mozilla.org/firefox/" });
-    return false;
-  }
-
-  if (isCurrentTabValidUrlScheme) {
-    if (launchDefaultBrowsing) {
-      chrome.tabs.update(tab.id, { url: "firefox:" + tab.url });
-    } else {
-      chrome.tabs.update(tab.id, { url: "firefox-private:" + tab.url });
-    }
-    return true;
-  }
-  return false;
-}
-
-
-// Testing purposes
-export const setIsCurrentTabValidUrlScheme = (value) => {
-  isCurrentTabValidUrlScheme = value;
-};
