+++ src/shared/backgroundScripts/validTab.js
@@ -1,0 +1,11 @@
+export let isCurrentTabValidUrlScheme = false;
+
+export function checkAndUpdateURLScheme(tab) {
+  if (tab.url === undefined) isCurrentTabValidUrlScheme = false;
+  else if (tab.url.startsWith("http") || tab.url.startsWith("file")) {
+    isCurrentTabValidUrlScheme = true;
+  } else {
+    isCurrentTabValidUrlScheme = false;
+  }
+}
+
